Compute rtc parameters when we assemble rtc
