+++ dotcom-rendering/src/amp/lib/real-time-config.ts
@@ -18,32 +18,32 @@ export type AdType =
  * These can be computed from the Config type above
  */
 export type RTCParameters = {
-	profileId: string;
-	pubId: string;
+	PROFILE_ID: string;
+	PUB_ID: string;
 };
 
 /**
  * Determine the pub id and profile id required by Pubmatic to construct an RTC vendor
  *
  */
-export const getRTCParameters = (adType: AdType): RTCParameters => {
+export const pubmaticRtcParameters = (adType: AdType): RTCParameters => {
 	if (
 		adType.isSticky ||
 		adType.adRegion === 'UK' ||
 		adType.adRegion === 'INT'
 	) {
 		return {
-			profileId: '6611',
-			pubId: '157207',
+			PROFILE_ID: '6611',
+			PUB_ID: '157207',
 		};
 	}
 
 	if (adType.adRegion === 'AU') {
-		return { profileId: '6697', pubId: '157203' };
+		return { PROFILE_ID: '6697', PUB_ID: '157203' };
 	}
 
 	// ad region is US
-	return { profileId: '6696', pubId: '157206' };
+	return { PROFILE_ID: '6696', PUB_ID: '157206' };
 };
 
 const permutiveURL = 'amp-script:permutiveCachedTargeting.ct';
@@ -52,31 +52,24 @@ const amazonConfig = {
 	aps: { PUB_ID: '3722', PARAMS: { amp: '1' } },
 };
 
-const notUndefined = <T>(x: T | undefined): x is T => x !== undefined;
-
 /**
  * Build a generic Real Time Config string from a possible URL,
  * optional vendors and whether to enable Permutive and Amazon
  */
-export const realTimeConfig = ({
-	vendors = {},
-	url = undefined,
-	usePermutive = false,
-	useAmazon = false,
-}: {
-	vendors?: Record<string, unknown>;
-	url?: string;
-	usePermutive?: boolean;
-	useAmazon?: boolean;
-}): string => {
-	const urls = [url, usePermutive ? permutiveURL : undefined].filter(
-		notUndefined,
-	);
+export const realTimeConfig = (
+	usePrebid: boolean,
+	usePermutive: boolean,
+	useAmazon: boolean,
+	adType: AdType,
+): string => {
+	const pubmaticConfig = {
+		openwrap: pubmaticRtcParameters(adType),
+	};
 
 	const data = {
-		urls,
+		urls: usePermutive ? [permutiveURL] : [],
 		vendors: {
-			...vendors,
+			...(usePrebid ? pubmaticConfig : {}),
 			...(useAmazon ? amazonConfig : {}),
 		},
 		timeoutMillis: 1000,
